refactor(validation): optimize tests to improve code test coverage
